отчистка проекта от лишних файлов и упорядочивание кода
