+++ FakeNFT/Scenes /Catalog/CollectionScreen/CollectionViewController.swift
@@ -119,8 +119,6 @@ final class CollectionViewController: UIViewController, ErrorView {
     
     private var collectionCellWidth: CGFloat = 0
     
-   // private var collectionOfMockNft: [MockNftStatistics] = []
-    
     private var topBarHeight: CGFloat {
         let statusBarHeight: CGFloat = 54.0
         let navBarHeight: CGFloat = 96
@@ -138,21 +136,16 @@ final class CollectionViewController: UIViewController, ErrorView {
     }
     
     //MARK: - viewDidLoad
-    
     override func viewDidLoad() {
         super.viewDidLoad()
-        print(presenter.nftModel)
-     //   setMockData()
         setupPresenter()
         setupUI()
     }
     
+    //MARK: - public funcs
+    
     func setupPresenter() {
-     //   presenter.getLikesAndCart()
-       
         presenter.getLikesCartAndNft()
-//        presenter.loadAuthorWebsite()
-//        presenter.prepareFullDataForShow()
     }
     
     func show(viewCollectionViewModel model: CollectionViewModel) {
@@ -168,9 +161,7 @@ final class CollectionViewController: UIViewController, ErrorView {
         nftCollectionView.reloadData()
     }
     
-    func updateData(with nfts: [Nft], id: String?, isCart: Bool?) {
-      //  nftCollectionView.reloadData()
-        
+    func reloadCollectionForLikesAndCard(with nfts: [Nft], id: String?, isCart: Bool?) {
         for (index, nft) in nfts.enumerated() {
             if let cell = nftCollectionView.cellForItem(at: IndexPath(item: index, section: 0)) as? NFTCollectionViewCellThreePerRow {
                 cell.set(data: nft)
@@ -195,29 +186,6 @@ final class CollectionViewController: UIViewController, ErrorView {
         navigationItem.leftBarButtonItem = backButton
     }
     
-    //MARK: - public funcs
-    
-//    func setMockData() {
-//        collectionOfMockNft = [
-//            MockNftStatistics(name: "Archie", images: [Asset.MockImages.Peach.Archie._1.image], rating: 2, price: 1, id: "1"),
-//            MockNftStatistics(name: "Art", images: [Asset.MockImages.Peach.Art._1.image], rating: 3, price: 2, id: "2"),
-//            MockNftStatistics(name: "Biscuit", images: [Asset.MockImages.Peach.Biscuit._1.image], rating: 1, price: 3, id: "3"),
-//            MockNftStatistics(name: "Daisy", images: [Asset.MockImages.Peach.Daisy._1.image], rating: 4, price: 13, id: "4"),
-//            MockNftStatistics(name: "Nacho", images: [Asset.MockImages.Peach.Nacho._1.image], rating: 5, price: 4, id: "5"),
-//            MockNftStatistics(name: "Oreo", images: [Asset.MockImages.Peach.Oreo._1.image], rating: 2, price: 2, id: "6"),
-//            MockNftStatistics(name: "Pixi", images: [Asset.MockImages.Peach.Pixi._1.image], rating: 1, price: 1, id: "7"),
-//            MockNftStatistics(name: "Ruby", images: [Asset.MockImages.Peach.Ruby._1.image], rating: 3, price: 3, id: "8"),
-//            MockNftStatistics(name: "Susan", images: [Asset.MockImages.Peach.Susan._1.image], rating: 2, price: 41, id: "9"),
-//            MockNftStatistics(name: "Tater", images: [Asset.MockImages.Peach.Tater._1.image], rating: 2, price: 2, id: "10")
-//        ]
-        
-//        coverImageView.image = Asset.MockImages.CollectionCovers.beige.image
-//        titleLabel.text = "Beige"
-//        authorLinkLabel.text = "Mock Link"
-//        collectionDescriptionLabel.text = "Персиковый — как облака над закатным солнцем в океане. В этой коллекции совмещены трогательная нежность и живая игривость сказочных зефирных зверей."
-        
-//    }
-    
     //MARK: - private funcs
     
     @objc private func authorLinkTapped() {
@@ -225,7 +193,7 @@ final class CollectionViewController: UIViewController, ErrorView {
     }
     
     @objc private func backButtonTapped() {
-        navigationController?.popViewController(animated: true)
+        presenter.backButtonTapped()
     }
     
     private func loadCoverImage(url: String) {
@@ -246,7 +214,13 @@ extension CollectionViewController: UICollectionViewDataSource {
 
         let cell: NFTCollectionViewCellThreePerRow = collectionView.dequeueReusableCell(indexPath: indexPath)
         cell.delegate = self.presenter
-        cell.set(data: presenter.nftArray[indexPath.row])
+        let nft = presenter.nftArray[indexPath.row]
+        cell.set(data: nft)
+        
+        let idOfCell = nft.id
+        cell.setLikedStateToLikeButton(isLiked: presenter.isLiked(idOfCell))
+        cell.setAddedStateToCart(isAdded: presenter.isAddedToCart(idOfCell))
+        
         return cell
     }
 }
