+++ FakeNFT/Services/Requests/ProfileRequest.swift
@@ -21,6 +21,6 @@ struct ProfileRequest: NetworkRequest {
     }
     
     var endpoint: URL? {
-        URL(string: "\(RequestConstants.baseURL)/api/v1/profile/1")
+        URL(string: "\(NetworkConstants.baseURL)/api/v1/profile/1")
     }
 }
