--- FakeNFT/Services/Requests/RequestConstants.swift
@@ -1,3 +1,0 @@
-enum RequestConstants {
-    static let baseURL = "https://d5dn3j2ouj72b0ejucbl.apigw.yandexcloud.net"
-}
